Merged PR 18955: engineering: Add clean_install_start metric to help indicate when failures in trident occur for PowerBI dashboard
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

Adding a clean_install_start metric in trident so when a deployment test fails for whatever reason, we can use KQL to consider it a failure to represent as a success/fail rate in PowerBI.

# 🤔 Rationale

Why is this PR needed?
Allows for success failure indication for dashboards, and filtering by pipeline (trident-ci, trident-prerelease, etc)

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to the software, specifically adding a `clean_install_start` metric to help track failures in the trident process for the PowerBI dashboard.

#### PR Summary
The changes in this pull request primarily involve modifications to the pipeline templates and the addition of a new metric in the trident process. These changes aim to improve the tracking of failures in the trident process and enhance the PowerBI dashboard's data representation.
- In `/.pipelines/templates/stages/deployment_testing/common_tasks/trident-metrics.yml`, a new parameter `runOnFailure` was added and conditions for running tasks were updated to include this parameter.
- In `/.pipelines/templates/stages/deployment_testing/vm/netlaunch-testing.yml`, the `runOnFailure` parameter was set to true.
- In `/e2e_tests/trident_configurations/base/trident-config.yaml`, a minor change was made to the `content` field.
- In `/src/modules/mod.rs`, a new metric `clean_install_start` was added to the `clean_install` function to track the start of the clean install process.

Related work items: #7709