Merged PR 19042: bug: Fix some issues with run-netlaunch-sample
# 🔍 Description

The `make run-netlaunch-sample` command was not updated when we refactored the storage API and some other issues have slipped into the sample configs.

# 📝 Checks

Locally tested the various configs (using  `make HOST_CONFIG=simple.yaml make run-netlaunch-sample`, etc.) to ensure that they passed the validation stage.

Notably, none of the configs actually succeeded in provisioning the VM...

- basic.yaml: Trident repeatedly crashed on the runtime OS (I probably need to update my local runtimeOS images to have newer Trident...)
- simple.yaml: Fails to due "/var/config-script.sh" not existing on the provOS
- base.yaml: Fails to due "/var/config-script.sh" not existing on the provOS
- verity.yaml: Hangs at boot, perhaps due to an overlay mount issue
- advanced.yaml: Hangs at boot, perhaps due to an overlay mount issue

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix, addressing issues with the `run-netlaunch-sample` in the Trident API.

#### PR Summary
The changes primarily involve modifications to the device paths and mount options in the host configuration samples. The FileSystem block for the 'srv' device has also been removed.
- In `trident_api/src/samples/sample_hc.rs`, the device paths have been updated and the FileSystem block for 'srv' has been removed.
- In `/docs/Reference/Host-Configuration/Samples/advanced.yaml`, the device paths have been updated and the FileSystem block for 'srv' has been removed.
- In `/Makefile`, the command for `run-netlaunch-sample` has been updated to ignore certain sha256 values.
- In various sample configuration files (e.g., `base.yaml`, `basic.yaml`, `simple.yaml`, `verity.yaml`, `sample-host-configuration.md`, `sample-host-configuration.yaml`), the device paths have been updated.

Related work items: #7432