Merged PR 19055: bug: Fix unit test of sample HC
# 🔍 Description

Forgot to update unit test when modifying the advanced sample config.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix, specifically addressing a failing unit test in the Trident CI pipeline.

#### PR Summary
The pull request modifies a unit test in the `sample_hc.rs` file to correct a failing test in the Trident CI pipeline.
- In `sample_hc.rs`, the expected number of filesystems in the `host_configuration.storage` object has been updated from 11 to 10. This change aligns the test with the current state of the software, resolving the test failure.

Related work items: #7749