Merged PR 19062: feature: TARGET_ROOT and EXEC_ROOT env variables for scripts
# 🔍 Description

This enables user scripts to know the paths to the target root (where the runtime OS is) and the exec root (where Trident is running from)

# 📌 Follow-ups

The TARGET_ROOT path defaults to /mnt/newroot when possible. Should we randomize it so that scripts don't hardcode that directory and then fail if Trident uses the fallback path?

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature that adds `TARGET_ROOT` and `EXEC_ROOT` as environment variables for scripts.

#### PR Summary
The changes in this pull request primarily modify the way scripts are run and environment variables are set. The `TARGET_ROOT` and `EXEC_ROOT` are now passed as environment variables to the scripts, providing more flexibility and control.
- In `src/modules/hooks.rs`, the `run_script` and `set_env_vars` functions have been refactored to include `servicing_type`, `target_root`, and `exec_root` as parameters instead of `host_status`.
- The `TARGET_ROOT` and `EXEC_ROOT` environment variables are now set in the `set_env_vars` function.
- The `run_script` function now checks if the script should run based on the `servicing_type` parameter.
- The tests have been updated to reflect these changes, with `host_status` replaced by the new parameters in the function calls.
- The `ROOT_MOUNT_POINT_PATH` constant is now used in the `run_script` function calls.

Related work items: #7751