Merged PR 19092: bug: Valid semver  trident version
# 🔍 Description

Adding a `v` at the beginning of the short commit for the trident version

# 🤔 Rationale

Currently, we are using for trident version the: `<version core> "-" <pre-release>` format, where:
```
<pre-release> ::= <dot-separated pre-release identifiers>
<dot-separated pre-release identifiers> ::= <pre-release identifier>
                                          | <pre-release identifier> "." <dot-separated pre-release identifiers>
```
```
<pre-release identifier> ::= <alphanumeric identifier>
                           | <numeric identifier>

```
```
<numeric identifier> ::= "0"
                       | <positive digit>
                       | <positive digit> <digits>
```
So, when the short commit doesn't have any letters (`<numeric identifier>`) and it start with `0` it won't be valid (invalid numeric identifier). Adding a `v` at the beginning solves the issue (always an alphanumeric identifier).

https://semver.org/spec/v2.0.0.html

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix, addressing an issue with the trident version not being a valid semver in the continuous integration process.

#### PR Summary
This PR modifies the versioning scheme in the `get-version.py` script to ensure the generated trident version is a valid semver. This change is expected to resolve the CI failure linked to this PR.
- In `/scripts/get-version.py`, the version string format has been updated to include a 'v' before the short commit hash when the `--commit` argument is used.

Related work items: #7763