Merged PR 19083: bug: Fix Functional Tests filter
# 🔍 Description

Fix `make functional-test` with `FILTER=` specified.

# 🗒️ Notes

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix aimed at improving the functionality of the test filter in the Makefile.

#### PR Summary
The pull request modifies the Makefile to allow for a subset of tests to be run while still triggering the VM deployment. This addresses the issue where the FILTER would prevent the deployment test from running.
- In the Makefile, the command for running functional tests has been updated to include `functional_tests/test_setup.py` before the `$(FILTER)`. This ensures that the test setup is always run, regardless of the filter applied.

Related work items: #7312