Merged PR 19142: engineering: Add negative test cases to trident_api/src/config/host/*
#### AI description  (iteration 1)
#### PR Classification
This pull request is focused on enhancing the test coverage by adding negative test cases to the `additional_files.rs` file.

#### PR Summary
The pull request introduces two new negative test cases to the `additional_files.rs` file to ensure the robustness of the code. These tests are designed to validate the failure scenarios of the `AdditionalFile` struct.
- In `/trident_api/src/config/host/os/additional_files.rs`, a test case `test_additional_files_validate_fail_both_content_path` has been added to validate the scenario where both `content` and `path` fields are present in the `AdditionalFile` struct.
- Another test case `test_additional_files_validate_fail_no_content_or_path` has been added in the same file to validate the scenario where both `content` and `path` fields are absent in the `AdditionalFile` struct.

Related work items: #7063