Merged PR 19134: feature: trident.service should not rerun on failure
# 🔍 Description

Have the Trident's systemd service not rerun if it fails.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature that modifies the behavior of the `trident.service` to prevent it from rerunning on failure.

#### PR Summary
The changes in this pull request aim to enhance the robustness of the `trident.service` by preventing it from rerunning automatically upon encountering a failure. This is achieved by modifying the service's configuration and adding a new subprocess run command in the `trident-config.yaml` file.
- In `/systemd/trident.service`, the `Restart=on-failure` and `RestartSec=5s` lines have been removed, disabling the automatic restart of the service on failure.
- In `/e2e_tests/trident_configurations/rerun/trident-config.yaml`, a new subprocess run command has been added which runs the `trident` command with `DEBUG` verbosity only if the file `/run/already-run` does not exist. This ensures that the service does not rerun on failure.

Related work items: #7063