Merged PR 19059: engineering: Implement span logic to use tracing macros for granular time measurements
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

Tracing comes with a lot of useful features for simplifying instrumentation. For getting timings for most of the function logic within Trident, we can use the `#[tracing::instrument()]` macro on any function we want to get an automatic span created and time measured. The execution time measured happens as the custom layer TraceSender implements the `on_entry` and `on_exit` which get called anytime a span is automatically closed (when function with macro finishes) or manually exited.

TODO:
- [X] Format request similar to `on_event` when a span exits to send to netlaunch.
- [X] Figure out how to handle spans that don't technically close automatically because of the `reboot()`

# 🤔 Rationale

Why is this PR needed?

Allows for faster instrumentation with less manual timing logic for each function and part of trident.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

Current breakdown:
![image (2).png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/19059/attachments/image%20%282%29.png)

----
#### AI description  (iteration 1)
#### PR Classification
This pull request implements span logic for granular time measurements in the Trident project, enhancing the tracing capabilities for better performance analysis.

#### PR Summary
The changes in this pull request introduce tracing spans for various functions across multiple modules, allowing for more precise time measurements. This is part of the work to provide a granular breakdown of Trident's operations.
- In `/src/logging/tracestream.rs`, new methods `on_enter` and `on_exit` were added to handle span creation and completion, respectively.
- In `/src/modules/mod.rs`, several functions were instrumented with tracing spans, including `clean_install`, `validate_host_config`, `reboot`, and others.
- In `/Makefile`, a new flag `-m trident-metrics.json` was added to the `run-netlaunch` command to output metrics.
- In various other modules like `/src/modules/storage/raid.rs`, `/src/modules/hooks.rs`, `/src/modules/storage/encryption.rs`, and others, specific functions were instrumented with tracing spans.
- In `/src/modules/bootentries.rs`, the `set_boot_order` and `update_efi_boot_order` functions were instrumented with tracing spans.

Related work items: #7722