Merged PR 19149: engineering: Create release pipeline
# 🔍 Description

Finally set up the release pipeline.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is focused on creating a release pipeline for the software project.

#### PR Summary
The changes in this pull request primarily involve the addition of new YAML files and modifications to existing ones to set up a release pipeline. The pipeline is designed to handle specific versions of packages and has the capability to download the latest version if no specific version is provided.
- In `/.pipelines/templates/stages/common_tasks/download-universal-artifact.yml`, a new parameter `specificVersion` was added to allow downloading of a specific package version. The script was also modified to handle this new parameter.
- A new file `/.pipelines/trident-release.yml` was added, which sets up the main structure of the release pipeline.
- Another new file `/.pipelines/templates/release-template.yml` was added, which serves as a template for the release pipeline.
- In `/.pipelines/templates/stages/trident_rpms/download-staged.yml`, the `specificVersion` parameter was added to the download stage of the pipeline.
- In `/.pipelines/templates/pipeline-selector.yml`, a minor change was made to remove the "rel" stage.