Merged PR 19164: engineering: Add absolute path validation for script files
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

Similar to block device paths, we expect absolute paths to any script files listed in the host config. This PR adds a validation check for each script that if a path is provided, it is an absolute path.

# 🤔 Rationale

Why is this PR needed?

Prevents common errors trying to refer to script files with relative path.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces new features to validate absolute paths for script and device files in the Trident API.

#### PR Summary
The changes in this pull request ensure that only absolute paths are used for script and device files, improving the robustness of the system. This is achieved by adding validation checks and corresponding test cases.
- In `scripts.rs` and `storage/mod.rs`, new test cases have been added to validate the use of absolute paths.
- In `storage/mod.rs` and `blkdev_graph/rules/mod.rs`, validation checks have been added to ensure that only absolute paths are used.
- In `error.rs`, a new error type `PathNotAbsolute` has been introduced to handle cases where a non-absolute path is used.
- In `storage/mod.rs`, the device path is ensured to start with '/dev' for valid block devices.

Related work items: #7063