Merged PR 19181: engineering: Show netlisten deployment logs
# 🔍 Description

Showing logs even when the previous task failed by grouping Display and publish task.

# 🤔 Rationale

We need to show the deployment logs specially when the previous task failed for easier debugging. To not replicate logic of publish task we grouped the two tasks together in one.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a code cleanup to improve the logging process during baremetal deployment testing.

#### PR Summary
The pull request modifies the `baremetal-deploy.yml` file to streamline the process of displaying and publishing logs collected from the netlisten process.
- In `/.pipelines/templates/stages/deployment_testing/baremetal/baremetal-deploy.yml`, the separate steps for displaying and publishing logs have been merged into a single step, reducing redundancy and improving code readability.

Related work items: #7839