Merged PR 19174: engineering: Quick fix to prevent bin directory being read-only when running netlaunch
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?
When you run `make run-netlaunch` after cloning trident or after a `make clean` it can cause the `bin/` directory to be read only causing it to fail the first run. This should fix that issue.

# 🤔 Rationale

Why is this PR needed?
Prevents failure on first run with trident and netlaunch.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

quick fix to prevent bin directory being read-only when running netlaunch

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a quick fix to prevent the bin directory from being read-only when running netlaunch.

#### PR Summary
The changes in this pull request primarily involve the addition of commands to create the 'bin' directory in the Makefile. This is aimed at preventing the directory from being read-only when running netlaunch.
- In the Makefile, a command to create the 'bin' directory was added before the execution of the 'imagecustomizer' in the 'trident-mos.iso' and 'trident-containerhost-mos.iso' targets.

Related work items: #7063