Merged PR 19214: engineering: Fix yaml file name
# 🔍 Description

Correct file extension.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request involves a code cleanup operation, specifically the removal of an entire yaml file.

#### PR Summary
The pull request removes the entire `/.pipelines/trident-prerelease.yaml` file from the codebase. This file contained configurations for parameters, schedules, triggers, variables, resources, and stages for the pipeline. The removal of this file may affect the pipeline's operation, and it's crucial to ensure that these configurations are not required elsewhere in the project.
- Removed the entire `/.pipelines/trident-prerelease.yaml` file.

Related work items: #7740