Merged PR 19213: engineering: Update values of ServicingState following the renaming of allowedOperations
# 🔍 Description

This PR updates the values of ServicingState following the renaming of allowedOperations to 'stage' and 'finalize'.

# 🤔 Rationale

This renaming is important for consistency since originally, the values of ServicingState were designed to correspond to the values of allowedOperations, e.g. StageDeployment -> StagingDeployment, DeploymentStaged; FinalizeDeployment -> FinalizingDeployment, DeploymentFinalized. Since the word "deployment" has been removed for clarity and brevity from allowedOperations, it makes sense to modify the values of ServicingState in the same way. The user can thus request to have allowedOperations performed on the needed servicing type and then track the progress of this servicing inside the host status. Moreover, the relationship between servicingType and servicingState is now clearer, e.g.:

`servicingType: clean-install\n
servicingState: staging`

or

`servicingType: ab-update\n
servicingState: finalizing`

----
#### AI description  (iteration 1)
#### PR Classification
This pull request involves a code refactoring to update the naming of `ServicingState` values in the Trident project.

#### PR Summary
The pull request includes changes to the naming convention of `ServicingState` values across multiple files in the Trident project. The changes aim to improve code readability and maintainability.
- In `trident_api/src/status.rs`, the `ServicingState` enum values were renamed, for example, `StagingDeployment` was changed to `Staging`, and `DeploymentStaged` was changed to `Staged`.
- In `/README.md`, the documentation was updated to reflect the new `ServicingState` naming convention.
- In `/dev-docs/allowed-operations.md` and `/dev-docs/diagrams/state-diagrams/`, the documentation and diagrams were updated to reflect the new `ServicingState` naming convention.
- In `/src/lib.rs` and `/src/modules/mod.rs`, the code was updated to use the new `ServicingState` naming convention.
- In `/src/datastore.rs`, the tests were updated to use the new `ServicingState` naming convention.

Related work items: #7793